% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dviewer.R
\name{dataGraph}
\alias{dataGraph}
\title{Create a html widget for the data viewer}
\usage{
dataGraph(data, group = FALSE, width = NULL, height = NULL)
}
\arguments{
\item{data}{json string or list}
}
\description{
Create a html widget for the data viewer
}

