% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dviewer.R
\name{demoData}
\alias{demoData}
\title{set up demo with data}
\usage{
demoData()
}
\description{
Copy loaded R test_data to the demo_data.js file.
}
\details{
Use this function to replace the demo_data.js inside the library, so that it can be used for debugging the javascript.
}

