% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dviewer.R
\name{sigmaOutput}
\alias{sigmaOutput}
\title{Standard functions to make the html widget work with shiny}
\usage{
sigmaOutput(outputId, width = "100\%", height = "400px")
}
\description{
Standard functions to make the html widget work with shiny
}

